Ensure ShortNoSwap is always uint16_t.
It turns out gcc on little endian machines didn't guarantee the type of
ShortNoSwap due to it being a macro that just returned its parameter. At
the same time, LongNoSwap and FloatNoSwap have been fixed.

+++ include/QF/qendian.h
@@ -103,9 +103,9 @@ float	_FloatNoSwap (float f);
 #define FloatSwap(l) _FloatSwap (l)
 #endif
 
-#define ShortNoSwap(l) (l)
-#define LongNoSwap(l) (l)
-#define FloatNoSwap(l) (l)
+#define ShortNoSwap(l) ((uint16_t) (l))
+#define LongNoSwap(l) ((uint32_t) (l))
+#define FloatNoSwap(l) ((float) (l))
 
 // NOTE: these /always/ read and write /little/ endian entities.
 struct QFile_s;
